add : Qna 기능 구현.

+++ src/main/java/com/picketlogia/picket/api/qna/repository/QnaRepository.java
@@ -3,6 +3,6 @@ package com.picketlogia.picket.api.qna.repository;
 import com.picketlogia.picket.api.qna.model.Qna;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QnaRepository extends JpaRepository<Qna, Integer> {
+public interface QnaRepository extends JpaRepository<Qna, Long> {
 
 }
